Use double for balance amount. Fix number format on new transaction Fix scroll on new transaction

+++ src/commonMain/kotlin/uz/forbusiness/finance/repository/AppRepository.kt
@@ -8,7 +8,7 @@ import java.util.*
 interface AppRepository {
     suspend fun insertNewTransaction(
         note: String,
-        amount: Float,
+        amount: Double,
         dateTime: Date,
         fromAccount: Account,
         toAccount: Account
